Change CSS styling of index.html homepage
Will contintue to beautify this practice proect, I'll be using
it as a experimental tool to play with and try new things. I'm
a bit sleepy and tired now so I'm just going to keep going through
the TOP project, otherwise I would do the rest of the pages in
more depth.